wikisportcars/requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.27
Werkzeug==2.0.3
logging==0.5.1.2